+++ app/forms/conversion/edit_project_form.rb
@@ -30,8 +30,6 @@ class Conversion::EditProjectForm < EditProjectForm
 
     return false unless valid?
 
-    assigned_to = assigned_to_regional_caseworker_team ? nil : project.assigned_to
-    assigned_at = assigned_to_regional_caseworker_team ? nil : project.assigned_at
     team = assigned_to_regional_caseworker_team ? "regional_casework_services" : project.team
 
     project.assign_attributes(
@@ -43,8 +41,8 @@ class Conversion::EditProjectForm < EditProjectForm
       directive_academy_order: directive_academy_order,
       two_requires_improvement: two_requires_improvement,
       team: team,
-      assigned_to: assigned_to,
-      assigned_at: assigned_at
+      assigned_to: project.assigned_to,
+      assigned_at: project.assigned_at
     )
 
     ActiveRecord::Base.transaction do